bug fix, unsigned number compare with signed number

+++ src/kernels/cpu/add.cpp
@@ -52,17 +52,17 @@ namespace ts {
         }
 
         template<typename T>
-        static inline void compute_run_scalar(const T *plhs, T scalar, T *pout, size_t count) {
+        static inline void compute_run_scalar(const T *plhs, T scalar, T *pout, int count) {
             // this is CPU operator, so just using memcpy
             if (pout != plhs) std::memcpy(pout, plhs, count * sizeof(T));
 
-            for (size_t i = 0; i < count; ++i) {
+            for (int i = 0; i < count; ++i) {
                 reduce_operator(pout[i], scalar);
             }
         }
 
 		template<>
-		inline void compute_run_scalar(const float *plhs, float scalar, float *pout, size_t count) {
+		inline void compute_run_scalar(const float *plhs, float scalar, float *pout, int count) {
 //#ifdef TS_USE_OPENMP
 //			//#pragma omp parallel for num_threads(1)
 //#pragma omp parallel for num_threads(openmp_threads(count))
@@ -78,17 +78,17 @@ namespace ts {
 		}
 
         template<typename T>
-        static inline void compute_run_same_shape(const T *plhs, const T *prhs, T *pout, size_t count) {
+        static inline void compute_run_same_shape(const T *plhs, const T *prhs, T *pout, int count) {
             // this is CPU operator, so just using memcpy
             if (pout != plhs) std::memcpy(pout, plhs, count * sizeof(T));
 
-            for (size_t i = 0; i < count;++i) {
+            for (int i = 0; i < count;++i) {
                 reduce_operator(pout[i], prhs[i]);
             }
         }
 
 		template<>
-		inline void compute_run_same_shape(const float *plhs, const float *prhs, float *pout, size_t count) {
+		inline void compute_run_same_shape(const float *plhs, const float *prhs, float *pout, int count) {
 //#ifdef TS_USE_OPENMP
 ////#pragma omp parallel for num_threads(1)
 //#pragma omp parallel for num_threads(openmp_threads(count))
@@ -111,7 +111,7 @@ namespace ts {
 
             auto scalar = prhs[0];
 
-            compute_run_scalar(plhs, scalar, pout, size_t(out.count()));
+            compute_run_scalar(plhs, scalar, pout, out.count());
         }
 
 
@@ -121,7 +121,7 @@ namespace ts {
             auto prhs = rhs.data<T>();
             auto pout = out.data<T>();
 
-            compute_run_same_shape(plhs, prhs, pout, size_t(out.count()));
+            compute_run_same_shape(plhs, prhs, pout, out.count());
         }
 
 
@@ -155,7 +155,7 @@ namespace ts {
                     for (int c = 0; c < channels; ++c) {
                         int offset = (n * channels + c) * count;
                         auto local_pout = pout + offset;
-                        compute_run_scalar(local_pout, prhs[channels], local_pout, size_t(count));
+                        compute_run_scalar(local_pout, prhs[channels], local_pout, count);
                     }
                 }
             }
